Extract organization-related models to their own module
